feat (model) : notebook for qunatization
