+++ .gitignore
@@ -12,3 +12,4 @@ params.json
 fine_tune_results/
 tb_logs
 logs
+*.pt
