+++ spoofing_detection_api/app/models/prediction_model.py
@@ -41,6 +41,7 @@ class SpoofDetector:
                 map_location=self.device,
             ),
         )
+        model = torch.compile(model, backend='inductor')
         model.eval()
         self.model = model
         return self.model
